pyproject.toml
[tool.poetry]
name = "sentimeterke"
version = "0.1.0"
description = "Twitter Sentiment Analysis using Python"
authors = ["Brian Koech <brightkoech@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.8.1"
tweepy = "^4.14.0"
numpy = "^1.24.3"
altair = "^5.0.1"
streamlit = "^1.23.1"
textblob = "^0.17.1"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
wordcloud = "^1.9.2"
pandas = "^2.0.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
